refactor(web): move dashboard stat helpers to apps/web/stats.py
Pure relocation of _get_export_statistics, _get_refresh_statistics, and
_get_forwarding_statistics from apps/web/views.py to a new apps/web/stats.py,
with no change to their behavior. The added per-period trend logic follows in
the next commit.
