+++ apps/web/tests/test_dashboard.py
@@ -20,7 +20,7 @@ from apps.refreshes.models import (
     RefreshConfig,
     RefreshRun,
 )
-from apps.web.views import _get_export_statistics, _get_forwarding_statistics, _get_refresh_statistics
+from apps.web.stats import _get_export_statistics, _get_forwarding_statistics, _get_refresh_statistics
 from tests.fixtures import authed_client, commcare_account, commcare_project, database
